Moved core/data to internal/core/data
Also changed previous app directory to be named core. My thinking is
that each tier of EdgeX (core, export, support) can have its own
internal directory rather than shoving everything into app. This
further reinforces encapsulation and scoping of the logic.
