+++ internal/core/command/Attribution.txt
