+++ internal/core/command/README.md
