+++ internal/core/command/const.go
