+++ internal/core/command/device.go
