+++ internal/core/command/rest.go
