+++ internal/core/data/Attribution.txt
