+++ internal/core/data/README.md
