+++ internal/core/data/const.go
