+++ internal/core/data/errors/bad-value-descriptor.go
