+++ internal/core/data/errors/unsupported-database-error.go
