+++ internal/core/data/event.go
