+++ internal/core/data/interfaces/db.go
