+++ internal/core/data/raml/core-data.raml
