+++ internal/core/data/reading.go
