+++ internal/core/data/utils.go
