+++ internal/core/data/validate.go
