+++ internal/core/data/validate_test.go
